+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/editor-mobile-bridge
 
+## 33.1.14
+
+### Patch Changes
+
+- Updated dependencies
+
+## 33.1.13
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 33.1.12
 
 ### Patch Changes